👌 move and rename sortNodesByTopologicalOrder
