+++ packages/rum-recorder/src/domain/rrweb/mutationObserver.spec.ts
@@ -2,7 +2,7 @@ import { isIE } from '@datadog/browser-core'
 import { createMutationPayloadValidator } from '../../../test/utils'
 import { snapshot, NodeType } from '../rrweb-snapshot'
 import { MutationController } from './mutation'
-import { startMutationObserver } from './mutationObserver'
+import { sortAddedAndMovedNodes, startMutationObserver } from './mutationObserver'
 import { MutationCallBack } from './types'
 
 describe('startMutationCollection', () => {
@@ -640,5 +640,52 @@ describe('startMutationCollection', () => {
         })
       })
     })
+  })
+})
+
+describe('sortAddedAndMovedNodes', () => {
+  let parent: Node
+  let a: Node
+  let aa: Node
+  let b: Node
+  let c: Node
+  let d: Node
+
+  beforeEach(() => {
+    // Create a tree like this:
+    //     parent
+    //     / | \ \
+    //    a  b c d
+    //    |
+    //    aa
+    a = document.createElement('a')
+    aa = document.createElement('aa')
+    b = document.createElement('b')
+    c = document.createElement('c')
+    d = document.createElement('d')
+    parent = document.createElement('parent')
+    parent.appendChild(a)
+    a.appendChild(aa)
+    parent.appendChild(b)
+    parent.appendChild(c)
+    parent.appendChild(d)
+  })
+
+  it('sorts siblings in reverse order', () => {
+    const nodes = [c, b, d, a]
+    sortAddedAndMovedNodes(nodes)
+    expect(nodes).toEqual([d, c, b, a])
+  })
+
+  it('sorts parents', () => {
+    const nodes = [a, parent, aa]
+    sortAddedAndMovedNodes(nodes)
+    expect(nodes).toEqual([parent, a, aa])
+  })
+
+  it('sorts parents first then siblings', () => {
+    const nodes = [c, aa, b, parent, d, a]
+    sortAddedAndMovedNodes(nodes)
+    expect(nodes).toEqual([parent, d, c, b, a, aa])
   })
 })
